+++ web/src/pages/OnboardingWizard.jsx
@@ -4,7 +4,7 @@ import {
   Sheet, Button, Stack, FormControl, FormLabel, Input, 
   Avatar, IconButton, Grid, Divider, CircularProgress,
   List, ListItem, ListItemButton, ListItemContent, ListItemDecorator,
-  Checkbox
+  Checkbox, Chip, Select, Option
 } from '@mui/joy';
 import { 
   Home, People, DirectionsCar, Inventory2,